Remove uint concept and use int instead

+++ NonogramSolver/Solver/Separators/Impl/WallSeparator.cs
@@ -5,7 +5,7 @@ namespace Solver.Engine.Separators.Impl
 {
     public class WallSeparator : ISeparator
     {
-        public Range[] Separate(List<uint> numbers, int[] fields)
+        public Range[] Separate(List<int> numbers, int[] fields)
         {
             Range[] ranges = new Range[numbers.Count];
 
@@ -14,15 +14,15 @@ namespace Solver.Engine.Separators.Impl
                 ranges[0] = new Range()
                 {
                     Start = 0,
-                    End = (uint)(numbers[0] - (fields.Length == numbers[0] ? 1 : 0))
+                    End = (numbers[0] - (fields.Length == numbers[0] ? 1 : 0))
                 };
             }
             if(numbers.Count >= 1 && fields[^1] == 1)
             {
                 ranges[numbers.Count - 1] = new Range()
                 {
-                    Start = (uint)(fields.Length - numbers[^1] - (fields.Length == numbers[^1] ? 0 : 1)),
-                    End = (uint)(fields.Length - 1)
+                    Start = (fields.Length - numbers[^1] - (fields.Length == numbers[^1] ? 0 : 1)),
+                    End = (fields.Length - 1)
                 };
             }
